drivers: correct parameters passed to strncmp
Many times when strncmp is called with the intent of ignoring the NULL terminator, the null terminator is accidentally included in that comparison, which in practice is just an strcmp with extra steps.

Subtract from the places where the intent seems to be to do a comparison without the NULL terminator.

Signed-off-by: Seija Kijin <doremylover123@gmail.com>