arch: change variable types to match those of what is assigned to them
There are many places where "int len = strlen(foo);" is written, just for len to be passed as a parameter of size_t. This causes truncation and then expansion back from int to size_t. Not to mention this is poor logic and needless truncations can add extra unneeded instructions.

This patch aims to fix the worst offenders.

Signed-off-by: Seija Kijin <doremylover123@gmail.com>