net: Fix for packets being rejected in the xHCI controller's ring buffer
When a packet larger than MTU arrives in Linux from the modem,
it is discarded with -EOVERFLOW error (Babble error).

This is seen on USB3.0 and USB2.0 busses.

This is because the MRU (Max Receive Size) is not a separate entity from the MTU (Max Transmit Size), and the received packets can be larger than those transmitted.

Following the babble error,
there was an endless supply of zero-length URBs
that were rejected with -EPROTO
(increasing the rx input error counter each time).

This is only seen on USB3.0.
These continue to come ad-infinitum
until the modem is shutdown.

There appears to be a bug in the core USB handling code
in Linux that doesn't deal with network MTUs smaller than 1500 bytes well.

By default, the dev->hard_mtu (the real MTU
is in lockstep with dev->rx_urb_size (essentially an MRU),
and the latter is causing trouble.

This has nothing to do with the modems,
as the issue can be reproduced by getting a USB-Ethernet dongle,
setting the MTU to 1430, and pinging with size greater than 1406.

Signed-off-by: Seija Kijin <doremylover123@gmail.com>
Co-Authored-By: TarAldarion <gildeap@tcd.ie>